Conflicts:
	copasi/CopasiDataModel/CCopasiDataModel.cpp
	copasi/UI/FunctionWidget1.cpp
	copasi/function/CExpression.cpp
	copasi/model/CModelValue.cpp
	copasi/parameterFitting/CExperiment.cpp
	copasi/plotUI/CopasiPlot.h
	copasi/sbml/CCellDesignerImporter.cpp
	copasi/sbml/SBMLImporter.cpp
	copasi/sbml/SBMLImporter.h
	copasi/steadystate/CNewtonMethod.cpp
	copasi/tssanalysis/CCSPMethod.cpp
	copasi/utilities/CCopasiMethod.cpp
	sbml-testsuite/wrapper.cpp

+++ copasi/UI/CQExpandModelData.cpp
@@ -1,4 +1,4 @@
-// Copyright (C) 2010 - 2013 by Pedro Mendes, Virginia Tech Intellectual
+// Copyright (C) 2010 - 2015 by Pedro Mendes, Virginia Tech Intellectual
 // Properties, Inc., University of Heidelberg, and The University
 // of Manchester.
 // All rights reserved.
@@ -119,7 +119,7 @@ void CQExpandModelData::slotOK()
           //check whether diffusion is requested for the metabolites inside
           size_t i;
 
-          for (i = 0; i < it->first->childCount(); ++i)
+          for (i = 0; i < (size_t)it->first->childCount(); ++i)
             {
               if (it->first->child(i)->checkState(1) == Qt::Checked)
                 {
